+++ HISTORY.rst
@@ -73,3 +73,5 @@
   used now in every app.
 - Add icons to the mark notification as seen and mark all notification as
   seen links.
+- Switch both view to mark notification as read to use the POST request
+  via a confirmation view.